# orchestrator-service/src/main/resources/openapi.yaml
openapi: 3.0.4
info:
  title: MaxQueue Orchestrator API
  description: API, which is the entry point to the MaxQueue project
  version: 0.0.1
servers:
  - url: http://orchestrator-service:8080
    description: Development server
  - url: http://localhost:8080
    description: Local server
...
paths:
  /users:
    post:
      tags: [users]
      summary: Add user in maxqueue system
      operationId: addUserInMaxQueueSystem
      parameters:
        - $ref: '#/components/parameters/AuthIdHeader'
        - $ref: '#/components/parameters/MaxHashHeader'
      requestBody:
        description: Add user in maxqueue system
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreatingRequest'
      responses:
        '200':
          description: Successful operation
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /users/{maxId}/mini-app:
    post:
      tags: [users]
      summary: Transfer data after opening the mini-app
      operationId: sendUserMiniAppData
      parameters:
        - $ref: '#/components/parameters/MaxIdPathParameter'
      requestBody:
        description: Transfer data after opening the mini-app
        content:
          application/json:
            schema:
              type: object
              properties:
                miniAppInitData:
                  type: string
              required:
                - miniAppInitData
      responses:
        '200':
          description: User was successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MiniAppInitResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
  /users/{maxId}:
    put:
      tags: [users]
      summary: Update user in maxqueue system
      operationId: updateUserInMaxQueueSystem
      parameters:
        - $ref: '#/components/parameters/MaxIdPathParameter'
        - $ref: '#/components/parameters/AuthIdHeader'
        - $ref: '#/components/parameters/MaxHashHeader'
      requestBody:
        description: Update user in maxqueue system
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Successful operation
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'

    get:
      tags: [users]
      summary: Get user by Max Messenger ID
      operationId: getUserByMaxId
      parameters:
        - $ref: '#/components/parameters/MaxIdPathParameter'
        - $ref: '#/components/parameters/AuthIdHeader'
        - $ref: '#/components/parameters/MaxHashHeader'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
  /organizations/{organizationId}/queues:
    post:
      tags: [organizations]
      summary: Create queue for organization
      operationId: createOrganizationQueue
      parameters:
        - $ref: '#/components/parameters/OrganizationIdPathParameter'
        - $ref: '#/components/parameters/AuthIdHeader'
        - $ref: '#/components/parameters/MaxHashHeader'
      requestBody:
        description: Add user in queue
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueueCreatingRequest'
      responses:
        '200':
          description: Queue created for organization
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'

    get:
      tags: [organizations]
      summary: Get list of queues in an organization
      operationId: getOrganizationQueues
      parameters:
        - $ref: '#/components/parameters/OrganizationIdPathParameter'
        - $ref: '#/components/parameters/AuthIdHeader'
        - $ref: '#/components/parameters/MaxHashHeader'
      responses:
        '200':
          description: Organization with its queues
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /organizations/{organizationId}/settings:
    get:
      tags: [organizations]
      summary: Get organization settings
      operationId: getOrganizationSettings
      parameters:
        - $ref: '#/components/parameters/OrganizationIdPathParameter'
        - $ref: '#/components/parameters/AuthIdHeader'
        - $ref: '#/components/parameters/MaxHashHeader'
      responses:
        '200':
          description: Organization settings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationSettingsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
...
  /queues/{queueId}/members:
    get:
      tags: [queues]
      summary: Get members of a queue
      operationId: getQueueMembers
      parameters:
        - $ref: '#/components/parameters/QueueIdPathParameter'
        - $ref: '#/components/parameters/AuthIdHeader'
        - $ref: '#/components/parameters/MaxHashHeader'
      responses:
        '200':
          description: Queue members retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueMembersResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'
  /queues/{queueId}/settings:
    get:
      tags: [queues]
      summary: Get queue settings
      operationId: getQueueSettings
      parameters:
        - $ref: '#/components/parameters/QueueIdPathParameter'
        - $ref: '#/components/parameters/AuthIdHeader'
        - $ref: '#/components/parameters/MaxHashHeader'
      responses:
        '200':
          description: Queue settings retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueSettingsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/UnexpectedError'

  #Analytics
...
components:
  parameters:
...
    MaxIdPathParameter:
      name: maxId
      in: path
      required: true
      description: Id provided by the Max massenger
      schema:
        type: integer
        format: int64
        example: 9932441
...
    OrganizationIdPathParameter:
      name: organizationId
      in: path
      required: true
      description: Organization id
      schema:
        type: string
        format: uuid
    QueueIdPathParameter:
      name: queueId
      in: path
      required: true
      description: Queue id
      schema:
        type: string
        format: uuid
...
    AuthIdHeader:
      name: Auth-Id
      in: header
      required: true
      description: Auth Id provided by orchestrator service
      schema:
        type: string
        format: uuid
    MaxHashHeader:
      name: Max-Hash
      in: header
      required: true
      description: Authentication hash provided by the Max massenger
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request parameters
    NotFound:
      description: Resource not found
    UnexpectedError:
      description: Unexpected server error
  schemas:
...
    UserRole:
      type: string
      description: Role of user
      enum:
        - CLIENT
        - EMPLOYEE
        - MODERATOR
        - ADMIN
    QueueEntryStatus:
      type: string
      description: Status of queue entry
      enum:
        - WAITING
        - SERVING
        - SERVED
        - CANCELED
        - MISSED

    ##### Frontend requests #####

    ### User Requests ###
    UserCreatingRequest:
      type: object
      properties:
        maxId:
          type: string
        username:
          type: string
        firstName:
          type: string
        secondName:
          type: string
      required:
        - maxId
      example:
        maxId: 123
        username: asd
        firstName: Alex
        secondName: Novikov
...
    UserUpdateRequest:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        secondName:
          type: string
        createdAt:
          type: string
          format: date-time
      example:
        username: asd
        firstName: Alex
        secondName: Novikov
        createdAt: 2025-11-09T16:21:09Z

    ### Queue Entry Request ###
...
    QueueCreatingRequest:
      type: object
      properties:
        name:
          type: string
        arrivalGracePeriod:
          type: integer
          format: int32
        maxQueueSize:
          type: integer
          format: in32
      required:
        - name
        - arrivalGracePreiod


    ##### Orchestrator responses #####

    ### Base Responses ###

    ### User Responses ###
    UserResponse:
      type: object
      properties:
        organizations:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        queue-entries:
          type: array
          items:
            $ref: '#/components/schemas/QueueEntryInUserResponse'
    QueueEntryInUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        peopleInFront:
          type: integer
        status:
          $ref: '#/components/schemas/QueueEntryStatus'
    MiniAppInitResponse:
      type: object
      properties:
        authId:
          type: string
          format: uuid
        maxHash:
          type: string
      required:
        - authId
        - maxHash

    ### Queue Entry Responses ###
...
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          allOf:
            - $ref: '#/components/schemas/UserRole'
            - enum: [ADMIN, MANAGER]
        amountOfQueues:
          type: integer
...
    OrganizationSettingsResponse:
      type: object
      properties:
        organization:
          $ref: '#/components/schemas/OrganizationSettings'
    OrganizationSettings:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        address:
          type: string
        isBanned:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # --- Organization metrics ---
...
    Queue:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Bakery Queue 1
        amountOfEmployees:
          type: integer
          format: int32
          example: 2
        maxSizeOfTodayQueue:
          type: integer
          format: int32
          example: 43
        amountOfServedPeople:
          type: integer
          format: int32
          example: 1024
    QueueResponse:
      properties:
        organizationId:
          type: string
          format: uuid
        organizationName:
          type: string
          example: Tasty Bakery
        queues:
          type: array
          items:
            $ref: '#/components/schemas/Queue'
    QueueSettingsResponse:
      properties:
        settings:
          $ref: '#/components/schemas/QueueSettings'
    QueueMembersResponse:
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/QueueMember'
...
    QueueSettings:
      type: object
      properties:
        arrivalGracePeriod:
          type: integer
        maxQueueSize:
          type: integer
        isActive:
          type: boolean

    # --- Queue metrics ---
...
    QueueMember:
      type: object
      properties:
        maxId:
          type: integer
          format: int64
        queueEntryId:
          type: string
          format: uuid
        username:
          type: string
...